aws_btc_price/requirements.txt
aws-psycopg2
requests
